# public-docs/openapi.yaml
openapi: 3.0.3
info:
  title: PKV Sync API
  version: 1.0.10
  license:
    name: AGPL-3.0-only
  contact:
    name: cyberkurry
  description: |
    HTTP API for PKV Sync clients and admin integrations.

    All `/api/*` routes are protected by the deployment key middleware and the
    plugin User-Agent guard in the production server stack. Authenticated user
    and admin routes additionally require a bearer device token. Production
    responses include clickjacking, MIME sniffing, referrer, and CSP headers;
    HSTS is emitted when `public_host` is configured.

    Starting at v1.0.0, this OpenAPI document is the public REST contract.
    Routes outside this document, including Admin Web UI form handlers, are
    internal implementation details unless documented separately.
servers:
  - url: https://sync.example.com
security:
  - DeploymentKey: []
    PluginUserAgent: []
...
components:
  securitySchemes:
    DeploymentKey:
      type: apiKey
      in: header
      name: X-PKVSync-Deployment-Key
      description: Deployment key distributed through the server share URL.
    PluginUserAgent:
      type: apiKey
      in: header
      name: User-Agent
      description: Must use the PKV Sync plugin User-Agent format accepted by the server.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: pks_token
...
  responses:
...
    Unauthorized:
      description: Missing, invalid, revoked, or expired bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    NotFound:
      description: Resource not found, or deployment key/User-Agent rejected by the production stack
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    TooManyRequests:
      description: Request rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
...
    JsonRpcRequest:
      type: object
      description: JSON-RPC 2.0 request object accepted by the MCP transport.
      additionalProperties: true
    JsonRpcResponse:
      type: object
      description: JSON-RPC 2.0 response object returned by the MCP transport.
      additionalProperties: true
paths:
...
  /mcp:
    post:
      tags: [MCP]
      security:
        - DeploymentKey: []
          BearerAuth: []
      summary: MCP Streamable HTTP JSON-RPC request
      description: |
        Streamable HTTP transport for Model Context Protocol clients. The
        endpoint is served by the optional MCP HTTP listener, not the main
        plugin REST listener. It requires the configured deployment key and a
        bearer device token. JSON-RPC method names and payloads follow the MCP
        protocol plus PKV Sync's documented tool set.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JsonRpcRequest'
      responses:
        '200':
          description: JSON-RPC response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JsonRpcResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
    get:
      tags: [MCP]
      security:
        - DeploymentKey: []
          BearerAuth: []
      summary: MCP Streamable HTTP SSE stream
      description: |
        Server-Sent Events stream for MCP vault change notifications. Clients
        must send `Accept: text/event-stream`; otherwise the server returns
        `405`. The stream uses bearer authentication, the deployment key, MCP
        replay via `Last-Event-ID`, and the same per-user/global SSE subscriber
        caps as the plugin event stream.
      parameters:
        - in: header
          name: Accept
          required: true
          schema:
            type: string
            enum: [text/event-stream]
        - in: header
          name: Last-Event-ID
          required: false
          schema:
            type: string
      responses:
        '200':
          description: MCP SSE notification stream
          content:
            text/event-stream:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '405':
          description: "Missing `Accept: text/event-stream`"
        '429':
          $ref: '#/components/responses/TooManyRequests'
...